test(pyreverse): split test_printer.py into per-printer test files
